LIBWEB-5393. Refactored code so DrupalGateway is a service
Refactored code to provide a "DrupalGateway" interface, with a
"DrupalGatewayImpl" class as the implementation of the interface.

Set up dependency injection to provide the UmdStaffDirectoryRestUpdater
class with the DrupalGateway implementation.

This is intended to make it easier to test the
UmdStaffDirectoryRestUpdater by using dependency injection to use a
mock class for testing.

Also modified code so that "UmdStaffDirectoryRestUpdater" is only
concerned with directory ids and JSON, and not any of the Drupal
internals. Drupal internals, such as node ids, or division taxonomy
ids are handled by the "DrupalGatewayImpl" class.

https://issues.umd.edu/browse/LIBWEB-5393