LIBWEB-5441. Only display "Advanced Search" for "WorldCat" search
Modified "hero_search" module so that the individual search targets
can optionally display an "alternate" search URL. The "Advanced Search"
link now only displays on the GUI when the "WorldCat" search is
selected.

Modified the settings so that the "search target" configuration is
specified using YAML. This allows an "alternate" URL to be specified
for search targets, which are then displayed when the associate
search target is selected.

Drupal "Conditional Form Fields" functionality
(https://www.drupal.org/docs/drupal-apis/form-api/conditional-form-fields)
is used to handle displaying/hiding the alternate search link.

https://issues.umd.edu/browse/LIBWEB-5441