LIBWEB-5462. Implemented News & Events View.
Pulls the 10 most recent articles, finds the first one with an image and puts it at the top of the list, then keeps the next 5. Displays via a grid with the featured article in the left third, the 5 other articles in the middle, and the events block in the right third.

This includes some basic styling for the layout and some of the text properties.

https://issues.umd.edu/browse/LIBWEB-5462