LIBWEB-5457. Added hero search target-configurable placeholder text
Modified the Hero search so that individual placeholder text can be
provided for search targets. Was implemented using the Drupal
"Conditional Form Fields" functionality, but creating individual
textfields for each search target, and then enabling only the relevant
textfield (based on the selected radio button).

Added handling for an (optional) "placeholder" key in the YAML
configuration for the search targets.

Changed the existing "Placeholder" configuration field with
"Default Placeholder", which will be used if no "placeholder" value is
provided from a search target.

Added a "hook_update" to "hero_search.install" to update an existing
configuration, replacing the "placeholder" field with
"default_placeholder" and configuring "placeholder" values for the
search targets.

https://issues.umd.edu/browse/LIBWEB-5457