LIBWEB-5429. Initial implementation of "ut_block" module
Added "ut_block" custom module to provide a "Block - 2 Columns"
paragraph type, allow custom blocks to be added to paragraphs using
the "block_field" modules.

The two column layout uses a CSS grid layout that is "left dominant",
the left column will take up at least 50% of the width with as much
additional width as necessary, leaving only the minimum amount needed
to show the content in the "right" column.

https://issues.umd.edu/browse/LIBWEB-5429