LIBWEB-5429. Improve and simplify newsroom CSS
Tweaked to newsroom CSS to provide for a display that takes better
advantage of available width.

Changed grid sizes to "auto", so that each column can take what it needs
in terms of sizing. This seems to be less wasteful of horizontal
whitespace.

Removed "events" related CSS, as the "UMD Terp Event" content type
should never be used in our implementation -- all events should
go through the "lib_cal" module.

https://issues.umd.edu/browse/LIBWEB-5429